- Moved crypto selection from radio buttons a dropdown which is easier to use. -  Added amount validation
