+++ src/validators/is-invalid-email.tsx
@@ -1,4 +1,4 @@
-import { areValidCcBcc } from './are-valid-cc-bcc';
+import { areValidCcBcc } from './are-valid-cc-bcc.tsx';
 
 export const isInvalidEmail = (email: string): boolean => {
   return !areValidCcBcc({ email });
