+++ src/validators/is-within-qr-size-limit.tsx
